Prevent overflow in index computation of fftshift

+++ idg-lib/src/fft/FFT.h
@@ -73,7 +73,7 @@ void fftshift(int m, int n, T *array) {
 template <typename T>
 void fftshift(int batch, int m, int n, T *array) {
   for (int i = 0; i < batch; i++) {
-    fftshift(m, n, &array[i * m * n]);
+    fftshift(m, n, &array[size_t(i) * size_t(m) * size_t(n)]);
   }
 }
